requirements.txt
scanpy
scvelo
torch
pyro-ppl
matplotlib
torchdiffeq